snowball: disabling -03 optimzation until we have new mali libs
The '-03' strips off '__aeabi_llsr' from libgcc.a, something that
prevents the mali libs from loading.  Those libs don't include the
symbol and expect an external entity to provide it.

This is only a temporary fix allowing snowball users to get a
working UI environment.  This should be reverted when we get new
mali libs, something we are currently working on.

Change-Id: Ifa4cc0b1671caa7a2cf5632b60b2aa5417b1a16c
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>